add os arquivos e updates do projeto - 28/08/23

+++ lib/src/pages/product/product_screen.dart
@@ -82,7 +82,7 @@ class _ProductScreenState extends State<ProductScreen> {
                             value: cartItemQuantity,
                             result: (quantity) {
                               setState(() {
-                                cartItemQuantity = quantity as int;
+                                cartItemQuantity = quantity;
                               });
                             },
                           ),
